+++ main.py
@@ -1,7 +1,7 @@
 '''
 Date: 2023-11-11 23:49:21
 LastEditors: Night-stars-1 nujj1042633805@gmail.com
-LastEditTime: 2023-11-13 18:21:45
+LastEditTime: 2023-11-13 20:14:46
 '''
 import asyncio
 
@@ -10,14 +10,14 @@ from utils.api.sign import BaseSign
 from utils.config import ConfigManager
 from utils.logger import log, get_message
 from utils.request import notify_me
-from utils.token import get_token_data
+from utils.utils import get_token
 
 _conf = ConfigManager.data_obj
 
 async def main():
     for account in _conf.accounts:
         login_obj = Login(account)
-        if (cookies := await login_obj.login()) and (token := await get_token_data(cookies["cUserId"])):
+        if (cookies := await login_obj.login()) and (token := await get_token(cookies["cUserId"])):
             sign_obj = BaseSign(cookies)
             daily_tasks = await sign_obj.check_daily_tasks()
             sign_task_obj = sign_obj.AVAILABLE_SIGNS # 签到任务对象合集
