+++ requirements.txt
@@ -4,3 +4,5 @@ loguru>=0.7.2
 pydantic>=2.4.2
 httpx>=0.25.1
 cryptography>=41.0.5
+pyyaml>=6.0.1
+tenacity>=8.2.3
